refactor(bank): Move database files to correct packages
