+++ Server/bank/database/src/main/java/com/cashmanager/server/database/enums/EnumAccountState.java
@@ -1,0 +1,6 @@
+package com.cashmanager.server.database.enums;
+
+public enum EnumAccountState {
+    ACTIVE,
+    INACTIVE
+}
