+++ Server/bank/database/src/main/java/com/cashmanager/server/database/enums/EnumLogSeverity.java
@@ -1,0 +1,7 @@
+package com.cashmanager.server.database.enums;
+
+public enum EnumLogSeverity {
+    INFO,
+    WARN,
+    ERROR
+}
