+++ Server/bank/database/src/main/java/com/cashmanager/server/database/enums/EnumPaymentMethodType.java
@@ -1,4 +1,4 @@
-package com.cashmanager.server.database.repositories;
+package com.cashmanager.server.database.enums;
 
 public enum EnumPaymentMethodType {
     CARD,
