+++ Server/bank/database/src/main/java/com/cashmanager/server/database/repositories/AccountRepository.java
@@ -1,5 +1,6 @@
-package com.cashmanager.server.database.repositories.account;
+package com.cashmanager.server.database.repositories;
 
+import com.cashmanager.server.database.entities.Account;
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.UUID;
