+++ Server/bank/database/src/main/java/com/cashmanager/server/database/repositories/AccountsLogsRepository.java
@@ -1,5 +1,6 @@
-package com.cashmanager.server.database.repositories.account;
+package com.cashmanager.server.database.repositories;
 
+import com.cashmanager.server.database.entities.AccountsLogs;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
