+++ Server/bank/database/src/main/java/com/cashmanager/server/database/repositories/TransactionLogRepository.java
@@ -1,5 +1,6 @@
-package com.cashmanager.server.database.repositories.transaction;
+package com.cashmanager.server.database.repositories;
 
+import com.cashmanager.server.database.entities.TransactionLog;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
