+++ Server/bank/database/src/main/java/com/cashmanager/server/database/repositories/TransactionRepository.java
@@ -1,5 +1,6 @@
-package com.cashmanager.server.database.repositories.transaction;
+package com.cashmanager.server.database.repositories;
 
+import com.cashmanager.server.database.entities.Transaction;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
