+++ Server/bank/database/src/main/java/com/cashmanager/server/database/repositories/UserRepository.java
@@ -1,5 +1,6 @@
-package com.cashmanager.server.database.repositories.user;
+package com.cashmanager.server.database.repositories;
 
+import com.cashmanager.server.database.entities.User;
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.UUID;
