--- Server/bank/database/src/main/java/com/cashmanager/server/database/repositories/account/EnumAccountState.java
@@ -1,6 +1,0 @@
-package com.cashmanager.server.database.repositories.account;
-
-public enum EnumAccountState {
-    ACTIVE,
-    INACTIVE
-}
