--- Server/bank/database/src/main/java/com/cashmanager/server/database/repositories/transaction/EnumLogSeverity.java
@@ -1,7 +1,0 @@
-package com.cashmanager.server.database.repositories.transaction;
-
-public enum EnumLogSeverity {
-    INFO,
-    WARN,
-    ERROR
-}
