+++ Server/bank/database/src/test/java/com/cashmanager/server/database/repositories/user/UserRepositoryTest.java
@@ -1,6 +1,9 @@
 package com.cashmanager.server.database.repositories.user;
 
 import com.cashmanager.server.database.DatabaseApplication;
+import com.cashmanager.server.database.entities.User;
+import com.cashmanager.server.database.enums.EnumUserRole;
+import com.cashmanager.server.database.repositories.UserRepository;
 import org.junit.jupiter.api.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
